Fix Select pins mapping.
Both the selector and demultiplexer are used to enable/disable the power
and data pins. If the mapping of pins is not identical between the 2
then we would enable the power pins of one USB port and the data pins of
another.

This patch changes that such that components which are closer to the
Pico are using the high-power bits while components which are further
down the tree are using the low bits. This way the enable pins are
correctly ordered as specified in the datasheets of both components.

Previously we were using the lowest bits, this would have been if
the enable pins were mapped accordingly. Which was not the case.

This change include changes in the schematics and in the board layout.
The component remain at their original placement.

This change also adds the OSHW logo as well as the fiducial in the
schematic such that they are not removed each time we re-import the
schematics in the board.