Add first draft of the firmware.
This firmware for the UF2 Batch Flasher provides a Web interface for the
Raspberry Pi Pico W.

The Web interface reports the status of the connected USB drives, and
combined with the web client, the image to be flashed can be repeated to
each device without adding cycles to the uf2 batch flasher flash.

The uf2 batch flasher listen for HTTP request to wait for instructions
and content to be flashed to each USB device. It act as a USB host,
powering one USB device at a time, which implies that the PC USB port
can power both the uf2 batch flasher and the usb device which is
flashed, thus not requiring any extra power supply.

As a USB host, it should request the board to switch to BOOTSEL mode,
without requiring the user to press the BOOTSEL button, nor the reboot
button, using a feature implemented as part of the as stdio_usb from the
Raspberry PI Pico default library, by selecting a BAUD rate of 1200.

Once the board is waiting for a flash image, the web client would
transmit the file over wifi, such that it can be flashed on the USB
device.

The Web client waits until all data have been transmitted before
switching to the next USB device to be flashed.

As a way to reduce the number of interfaces, the console is queued and
made available through the web server using stdout.txt file.