Properly report status via the ssi_handler.
The ssi_handler had some first draft issues, this is now fixed by
properly appending data and counting it properly. Also the stdout is now
properly processed by renaming the file with one of the extensions
supported by the SSI handler.

Testing the USB with another Pico, everything seems to work almost out
of the box, except that once switching the connected device to BOOTSEL
mode, the TinyUSB host does not unmount the device and stays in CDC.