firmware: Toggle USB data pins to move from CDC to MSC mount.
The Raspberry Pi Pico / RP2040, has a mode where we can set the BAUD rate at
1200Hz, to programmatically select the BOOTSEL mode, where it emulates a storage
device.

However TinyUSB does not take notice of this mode switch. Thus, we disable the
data pins until TinyUSB unmount the device from the CDC mode, and renable them
such that it can now consider it as an MSC mode. Such that we can flash the
device.