firmware: Stream bytes from POST to USB using pbuf from LwIP.
Previously the data was being copied over from the POST request to a pipe held
in memory, before being pulled by the USB host core which would send the bytes
over to the USB device. This techniques caused some packets drop as the pipe was
not being emptied fast enough and locked the receiving function.

Locking the receving function caused some packets drops which caused the flashed
image to be smaller than the original image which had to be flashed.

With these changed, the buffer is cloned to transmit the buffer to be used by
the USB host core, and then transmitted back to be freed as the pbuf allocator
is not thread safe. This approach might cause some problem in the future if the
RP2040 does not have enough RAM to hold the content to be flashed to the device,
but this seems to work so far with images of 127 KB.