firmware: Improve the asynchronous handling of state changes and flashed content.
This change refactor the code for checking the status code of each USB, such
that we can avoid repeating the same code for handling timeout, and we can focus
on the essential.

This code also fix the call to send_uf2, such that we properly call with the
loaded content instead of undefined.