web interface: Avoid failures, and allow reduced timeout.
Previously the hardware could stop by waiting for a status code which had
already been, then replaced. This implies that one would always wait for the
last stable status instead of waiting for the first status code.

By being able to wait for the last status code, we can increase the timer for
the last status code, while decreasing the timer for the first status code by
checking for inequality given that the most important status code are monotonic.