firmware: Wait to account for the reboot in BOOTSEL mode.
When we send the signal to put the device in BOOTSEL mode, by the time it
finishes its initialization, TinyUSB would have recorded the device as nothing
and then do not update its status after that.

By disabling the USB data lines, we force TinyUSB to ignore the intermediate
state and only consider the device when it is fully rebooted in BOOTSEL mode.
Thus, being ready to be mounted as a mass storage class device.