firmware: Do not f_sync after each f_write.
When dealing with a microcontroller and not a flash drive, we want to ensure
that each chunk is full of data that we intend to write.

f_sync call, while helping with flash drives, causes issues with microcontroller
devices being flashed with UF2 images. USB msc will typically write 512 bytes at
a time, except that the 512 bytes of UF2 represent a data structure which is
then interpreted by the microcontroller. Using f_sync will send partially
intialized data which has the potential of putting the microcontroller in a bad
state and potentially crashing it.

Thus this patch comments out the f_sync call within write_file_content to avoid
this issue.