firmware: Instrument USB host with disabled logging.
While investigating the f_sync issue, logging was necessary to figure out what
the issue was. However this instrumentation is costly both in space and time.

One of the issue while debugging was that the circular buffer of stdio_web was
sometimes losing content. This is solved by bumping the circular buffer to 32
KB. The time issue is solved in the web interface by increasing and splitting
the timeouts.