firmware: Add new state to avoid looping back.
When one flashing procedure is complete the USB device would restart after the
completion of te UF2 image. However, this causes the USB host to request that
the device is once more rebooted in BOOT select mode. The second BOOT select
mode would not be problematic if it were to not update the usb status, leaving
the system in a confused state.

This change introduced DEVICE_SELECTED, set once a device is selected and before
the DEVICE_BOOTSEL_REQUEST is reached. This way there is no loopback once the
content is properly flashed.

This change also clears all the USB states when starting the batch-flashing
procedure as a way to avoid loading stale content when flashing a second time.