Manage fetch calls to reduce pressure on the pico.
When flashing devices, the `/flash` request is copying the image in the RAM of
the flashing Pico, which does not leave a lot of margins for making additional
requests which would attempt to allocate and fail.

To avoid such failures, the automatic update of the status is disabled, an
additional asyncTimeout is added and the wait_for_status_update is now in charge
of spawning and killing the probing for status updates.

Additional mechanism added in the process, such as a mutex implementation, which
ensure that async operations are queued helps at preventing the overflow of
request while one is being processed.