Create a TCP server to send files using the command line.
Doing a drag and drop of file and click on a button is slower than using the
keyboard.

Also having a TCP server allow throttle down the quantity of data in the buffers
of the UF2 batch flasher, which has a limited amount of RAM compared to the
image size that can fit in the flash of the micro controllers.